Moved all tests from USSTreeBeforeRefactoring to abstract/TreeProvider.unit.test.ts
